[nxp][platform][common] Move Factory Data provider into a common folder
Some components were implemented for k32w1, but can be reused as long as the
used SDK is the same. Moving them to factory_data/legacy until synchronizing
with the other factory data implementation.
